Strip out as many dependncies from Lua Manager as possible to save memory
