+++ files/usr/local/bin/nvram-setup
@@ -72,7 +72,7 @@ if wifi_mac == "" or mac2 == "" then
 		if phy then
 			break
 		end
-		sleep(5);
+		sleep(5)
 	end
     for line in io.lines("/sys/class/ieee80211/" .. phy .. "/macaddress")
     do
